refactor: remove legacy loader pipeline
