+++ mcp_plex/loader/__init__.py
@@ -1,12 +1,4 @@
-"""Loader orchestration utilities and legacy pipeline helpers.
-
-This package now centres its public API on :class:`LoaderOrchestrator`,
-the coordination layer for the ingestion, enrichment, and persistence
-stages. The historical pipeline implementation remains available as
-``LoaderPipeline`` for reference, but new code should instantiate the
-staged classes directly and coordinate them with
-:class:`LoaderOrchestrator`.
-"""
+"""Loader orchestration utilities and staged pipeline helpers."""
 from __future__ import annotations
 
 import asyncio
@@ -36,7 +28,6 @@ from .imdb_cache import IMDbCache
 from .pipeline.channels import (
     IMDbRetryQueue,
     INGEST_DONE,
-    PERSIST_DONE,
     IngestBatch,
     IngestQueue,
     MovieBatch,
@@ -48,14 +39,20 @@ from .pipeline.channels import (
 )
 from .pipeline.orchestrator import LoaderOrchestrator
 from .pipeline.persistence import PersistenceStage as _PersistenceStage
+from .pipeline.enrichment import (
+    _build_plex_item,
+    _extract_external_ids,
+    _fetch_tmdb_episode,
+    _fetch_tmdb_movie,
+    _fetch_tmdb_show,
+    resolve_tmdb_season_number,
+)
 from ..common.types import (
     AggregatedItem,
-    ExternalIDs,
     IMDbTitle,
     PlexGuid,
     PlexItem,
     PlexPerson,
-    TMDBEpisode,
     TMDBItem,
     TMDBMovie,
     TMDBShow,
@@ -413,199 +410,6 @@ async def _ensure_collection(
     await _create_index("data.tmdb.id", models.PayloadSchemaType.INTEGER)
 
 
-async def _fetch_tmdb_movie(
-    client: httpx.AsyncClient, tmdb_id: str, api_key: str
-) -> Optional[TMDBMovie]:
-    url = (
-        f"https://api.themoviedb.org/3/movie/{tmdb_id}?append_to_response=reviews"
-    )
-    try:
-        resp = await client.get(url, headers={"Authorization": f"Bearer {api_key}"})
-    except httpx.HTTPError:
-        logger.exception("HTTP error fetching TMDb movie %s", tmdb_id)
-        return None
-    if resp.is_success:
-        return TMDBMovie.model_validate(resp.json())
-    return None
-
-
-async def _fetch_tmdb_show(
-    client: httpx.AsyncClient, tmdb_id: str, api_key: str
-) -> Optional[TMDBShow]:
-    url = (
-        f"https://api.themoviedb.org/3/tv/{tmdb_id}?append_to_response=reviews"
-    )
-    try:
-        resp = await client.get(url, headers={"Authorization": f"Bearer {api_key}"})
-    except httpx.HTTPError:
-        logger.exception("HTTP error fetching TMDb show %s", tmdb_id)
-        return None
-    if resp.is_success:
-        return TMDBShow.model_validate(resp.json())
-    return None
-
-
-async def _fetch_tmdb_episode(
-    client: httpx.AsyncClient,
-    show_id: int,
-    season_number: int,
-    episode_number: int,
-    api_key: str,
-) -> Optional[TMDBEpisode]:
-    """Fetch TMDb data for a TV episode."""
-
-    url = (
-        f"https://api.themoviedb.org/3/tv/{show_id}/season/{season_number}/episode/{episode_number}"
-    )
-    try:
-        resp = await client.get(url, headers={"Authorization": f"Bearer {api_key}"})
-    except httpx.HTTPError:
-        logger.exception(
-            "HTTP error fetching TMDb episode %s S%sE%s",
-            show_id,
-            season_number,
-            episode_number,
-        )
-        return None
-    if resp.is_success:
-        return TMDBEpisode.model_validate(resp.json())
-    return None
-
-
-def resolve_tmdb_season_number(
-    show_tmdb: Optional[TMDBShow], episode: PlexPartialObject
-) -> Optional[int]:
-    """Map a Plex episode to the appropriate TMDb season number.
-
-    This resolves cases where Plex uses year-based season indices that do not
-    match TMDb's sequential ``season_number`` values.
-    """
-
-    parent_index = getattr(episode, "parentIndex", None)
-    parent_title = getattr(episode, "parentTitle", None)
-    parent_year = getattr(episode, "parentYear", None)
-    if parent_year is None:
-        parent_year = getattr(episode, "year", None)
-
-    seasons = getattr(show_tmdb, "seasons", []) if show_tmdb else []
-
-    # direct numeric match
-    if parent_index is not None:
-        for season in seasons:
-            if season.season_number == parent_index:
-                return season.season_number
-
-    # match by season name (e.g. "Season 2018" -> "2018")
-    title_norm: Optional[str] = None
-    if isinstance(parent_title, str):
-        title_norm = parent_title.lower().lstrip("season ").strip()
-        for season in seasons:
-            name_norm = (season.name or "").lower().lstrip("season ").strip()
-            if name_norm == title_norm:
-                return season.season_number
-
-    # match by air date year when Plex uses year-based seasons
-    year: Optional[int] = None
-    if isinstance(parent_year, int):
-        year = parent_year
-    elif isinstance(parent_index, int):
-        year = parent_index
-    elif title_norm and title_norm.isdigit():
-        year = int(title_norm)
-
-    if year is not None:
-        for season in seasons:
-            air = getattr(season, "air_date", None)
-            if isinstance(air, str) and len(air) >= 4 and air[:4].isdigit():
-                if int(air[:4]) == year:
-                    return season.season_number
-
-    if isinstance(parent_index, int):
-        return parent_index
-    if isinstance(parent_index, str) and parent_index.isdigit():
-        return int(parent_index)
-    if isinstance(parent_title, str) and parent_title.isdigit():
-        return int(parent_title)
-    return None
-
-
-def _extract_external_ids(item: PlexPartialObject) -> ExternalIDs:
-    """Extract IMDb and TMDb IDs from a Plex object."""
-
-    imdb_id: Optional[str] = None
-    tmdb_id: Optional[str] = None
-    for guid in getattr(item, "guids", []) or []:
-        gid = getattr(guid, "id", "")
-        if gid.startswith("imdb://"):
-            imdb_id = gid.split("imdb://", 1)[1]
-        elif gid.startswith("tmdb://"):
-            tmdb_id = gid.split("tmdb://", 1)[1]
-    return ExternalIDs(imdb=imdb_id, tmdb=tmdb_id)
-
-
-def _build_plex_item(item: PlexPartialObject) -> PlexItem:
-    """Convert a Plex object into the internal :class:`PlexItem`."""
-
-    guids = [PlexGuid(id=g.id) for g in getattr(item, "guids", [])]
-    directors = [
-        PlexPerson(id=getattr(d, "id", 0), tag=str(getattr(d, "tag", "")), thumb=getattr(d, "thumb", None))
-        for d in getattr(item, "directors", []) or []
-    ]
-    writers = [
-        PlexPerson(id=getattr(w, "id", 0), tag=str(getattr(w, "tag", "")), thumb=getattr(w, "thumb", None))
-        for w in getattr(item, "writers", []) or []
-    ]
-    actors = [
-        PlexPerson(
-            id=getattr(a, "id", 0),
-            tag=str(getattr(a, "tag", "")),
-            thumb=getattr(a, "thumb", None),
-            role=getattr(a, "role", None),
-        )
-        for a in getattr(item, "actors", []) or getattr(item, "roles", []) or []
-    ]
-    genres = [
-        str(getattr(g, "tag", ""))
-        for g in getattr(item, "genres", []) or []
-        if getattr(g, "tag", None)
-    ]
-    collections = [
-        str(getattr(c, "tag", ""))
-        for c in getattr(item, "collections", []) or []
-        if getattr(c, "tag", None)
-    ]
-    season_number = getattr(item, "parentIndex", None)
-    if isinstance(season_number, str):
-        season_number = int(season_number) if season_number.isdigit() else None
-    episode_number = getattr(item, "index", None)
-    if isinstance(episode_number, str):
-        episode_number = int(episode_number) if episode_number.isdigit() else None
-
-    return PlexItem(
-        rating_key=str(getattr(item, "ratingKey", "")),
-        guid=str(getattr(item, "guid", "")),
-        type=str(getattr(item, "type", "")),
-        title=str(getattr(item, "title", "")),
-        show_title=getattr(item, "grandparentTitle", None),
-        season_title=getattr(item, "parentTitle", None),
-        season_number=season_number,
-        episode_number=episode_number,
-        summary=getattr(item, "summary", None),
-        year=getattr(item, "year", None),
-        added_at=getattr(item, "addedAt", None),
-        guids=guids,
-        thumb=getattr(item, "thumb", None),
-        art=getattr(item, "art", None),
-        tagline=getattr(item, "tagline", None),
-        content_rating=getattr(item, "contentRating", None),
-        directors=directors,
-        writers=writers,
-        actors=actors,
-        genres=genres,
-        collections=collections,
-    )
-
-
 def _format_primary_title(item: AggregatedItem) -> str:
     """Format the primary title text for ``item``."""
 
@@ -932,419 +736,6 @@ async def _iter_from_sample(sample_dir: Path) -> AsyncIterator[AggregatedItem]:
     for item in _load_from_sample(sample_dir):
         yield item
 
-
-class LoaderPipeline:
-    """Coordinate ingestion, enrichment, and Qdrant upserts."""
-
-    def __init__(
-        self,
-        *,
-        client: AsyncQdrantClient,
-        collection_name: str,
-        dense_model_name: str,
-        sparse_model_name: str,
-        tmdb_api_key: str | None,
-        sample_items: list[AggregatedItem] | None,
-        plex_server: PlexServer | None,
-        plex_chunk_size: int,
-        enrichment_batch_size: int,
-        enrichment_workers: int,
-        upsert_buffer_size: int,
-        max_concurrent_upserts: int,
-    ) -> None:
-        self._client = client
-        self._collection_name = collection_name
-        self._dense_model_name = dense_model_name
-        self._sparse_model_name = sparse_model_name
-        self._tmdb_api_key = tmdb_api_key
-        self._sample_items = sample_items
-        self._server = plex_server
-        self._plex_chunk_size = _require_positive(plex_chunk_size, name="plex_chunk_size")
-        self._enrichment_batch_size = _require_positive(
-            enrichment_batch_size, name="enrichment_batch_size"
-        )
-        self._enrichment_workers = _require_positive(
-            enrichment_workers, name="enrichment_workers"
-        )
-        self._upsert_buffer_size = _require_positive(
-            upsert_buffer_size, name="upsert_buffer_size"
-        )
-        self._max_concurrent_upserts = _require_positive(
-            max_concurrent_upserts, name="max_concurrent_upserts"
-        )
-
-        if self._sample_items is None and self._server is None:
-            raise RuntimeError("Either sample_items or plex_server must be provided")
-        if self._sample_items is None and not self._tmdb_api_key:
-            raise RuntimeError("TMDB API key required for live ingestion")
-
-        self._ingest_queue: IngestQueue = asyncio.Queue(
-            maxsize=self._enrichment_workers * 2
-        )
-        self._points_queue: PersistenceQueue = asyncio.Queue()
-        self._upsert_capacity = asyncio.Semaphore(self._max_concurrent_upserts)
-        self._items: list[AggregatedItem] = []
-        self._qdrant_retry_queue: asyncio.Queue[list[models.PointStruct]] = (
-            asyncio.Queue()
-        )
-        self._show_tmdb_cache: dict[str, TMDBShow | None] = {}
-
-        self._ingested_count = 0
-        self._enriched_count = 0
-        self._upserted_points = 0
-        now = time.perf_counter()
-        self._ingest_start = now
-        self._enrich_start = now
-        self._upsert_start = now
-        self._persistence_stage = _PersistenceStage(
-            client=self._client,
-            collection_name=self._collection_name,
-            dense_vector_name=self._dense_model_name,
-            sparse_vector_name=self._sparse_model_name,
-            persistence_queue=self._points_queue,
-            retry_queue=self._qdrant_retry_queue,
-            upsert_semaphore=self._upsert_capacity,
-            upsert_buffer_size=self._upsert_buffer_size,
-            upsert_fn=self._perform_upsert,
-            on_batch_complete=self._handle_upsert_batch,
-            worker_count=self._max_concurrent_upserts,
-        )
-
-    @property
-    def qdrant_retry_queue(self) -> asyncio.Queue[list[models.PointStruct]]:
-        """Expose the Qdrant retry queue for post-processing."""
-
-        return self._qdrant_retry_queue
-
-    @property
-    def items(self) -> list[AggregatedItem]:
-        """Return the aggregated items processed by the pipeline."""
-
-        return self._items
-
-    async def execute(self) -> None:
-        """Run the full ingestion/enrichment/upsert pipeline."""
-
-        async with httpx.AsyncClient(timeout=30) as client:
-            self._http_client = client
-            ingest_task = asyncio.create_task(self._ingest())
-            enrichment_tasks = [
-                asyncio.create_task(self._enrichment_worker(worker_id))
-                for worker_id in range(self._enrichment_workers)
-            ]
-            upsert_tasks = [
-                asyncio.create_task(self._persistence_stage.run(worker_id))
-                for worker_id in range(self._max_concurrent_upserts)
-            ]
-            error: BaseException | None = None
-            try:
-                await ingest_task
-                await self._ingest_queue.join()
-                await asyncio.gather(*enrichment_tasks)
-                await self._points_queue.join()
-            except BaseException as exc:
-                error = exc
-            finally:
-                for _ in range(self._max_concurrent_upserts):
-                    await self._points_queue.put(PERSIST_DONE)
-                upsert_results = await asyncio.gather(
-                    *upsert_tasks, return_exceptions=True
-                )
-                for result in upsert_results:
-                    if isinstance(result, BaseException) and not isinstance(
-                        result, asyncio.CancelledError
-                    ):
-                        if error is None:
-                            error = result
-                        break
-                for task in enrichment_tasks:
-                    if not task.done():
-                        task.cancel()
-                await asyncio.gather(*enrichment_tasks, return_exceptions=True)
-                if not ingest_task.done():
-                    ingest_task.cancel()
-                    await asyncio.gather(ingest_task, return_exceptions=True)
-                if error is not None:
-                    raise error
-
-    def _log_progress(
-        self, stage: str, count: int, start: float, queue_size: int
-    ) -> None:
-        elapsed = time.perf_counter() - start
-        rate = count / elapsed if elapsed > 0 else 0.0
-        logger.info(
-            "%s processed %d items (%.2f items/sec, queue size=%d)",
-            stage,
-            count,
-            rate,
-            queue_size,
-        )
-
-    async def _ingest(self) -> None:
-        start = time.perf_counter()
-        self._ingest_start = start
-        try:
-            if self._sample_items is not None:
-                await self._ingest_sample()
-            else:
-                await self._ingest_from_plex()
-        finally:
-            for _ in range(self._enrichment_workers):
-                await self._ingest_queue.put(None)
-            self._log_progress("Ingestion", self._ingested_count, start, self._ingest_queue.qsize())
-
-    async def _ingest_sample(self) -> None:
-        for chunk in _chunk_sequence(self._sample_items or [], self._enrichment_batch_size):
-            batch = _SampleBatch(items=list(chunk))
-            if not batch.items:
-                continue
-            await self._ingest_queue.put(batch)
-            self._ingested_count += len(batch.items)
-            self._log_progress(
-                "Ingestion",
-                self._ingested_count,
-                self._ingest_start,
-                self._ingest_queue.qsize(),
-            )
-
-    async def _ingest_from_plex(self) -> None:
-        if self._server is None:
-            raise RuntimeError("Plex server unavailable for ingestion")
-        movie_section = self._server.library.section("Movies")
-        movie_keys = [int(m.ratingKey) for m in movie_section.all()]
-        for key_chunk in _chunk_sequence(movie_keys, self._plex_chunk_size):
-            key_list = list(key_chunk)
-            movies = list(self._server.fetchItems(key_list)) if key_list else []
-            if not movies:
-                continue
-            await self._ingest_queue.put(_MovieBatch(movies=movies))
-            self._ingested_count += len(movies)
-            self._log_progress(
-                "Ingestion",
-                self._ingested_count,
-                self._ingest_start,
-                self._ingest_queue.qsize(),
-            )
-        show_section = self._server.library.section("TV Shows")
-        show_keys = [int(s.ratingKey) for s in show_section.all()]
-        for show_chunk in _chunk_sequence(show_keys, self._plex_chunk_size):
-            shows = list(self._server.fetchItems(list(show_chunk)))
-            for show in shows:
-                episode_keys = [int(e.ratingKey) for e in show.episodes()]
-                for episode_chunk in _chunk_sequence(episode_keys, self._plex_chunk_size):
-                    keys = list(episode_chunk)
-                    episodes = list(self._server.fetchItems(keys)) if keys else []
-                    if not episodes:
-                        continue
-                    await self._ingest_queue.put(
-                        _EpisodeBatch(show=show, episodes=episodes)
-                    )
-                    self._ingested_count += len(episodes)
-                    self._log_progress(
-                        "Ingestion",
-                        self._ingested_count,
-                        self._ingest_start,
-                        self._ingest_queue.qsize(),
-                    )
-
-    async def _enrichment_worker(self, worker_id: int) -> None:
-        while True:
-            batch = await self._ingest_queue.get()
-            if batch is None or batch is INGEST_DONE:
-                self._ingest_queue.task_done()
-                break
-            try:
-                if isinstance(batch, _MovieBatch):
-                    await self._process_movie_batch(batch)
-                elif isinstance(batch, _EpisodeBatch):
-                    await self._process_episode_batch(batch)
-                else:
-                    await self._process_sample_batch(batch)
-            finally:
-                self._ingest_queue.task_done()
-
-    async def _process_movie_batch(self, batch: _MovieBatch) -> None:
-        for chunk in _chunk_sequence(batch.movies, self._enrichment_batch_size):
-            movies = list(chunk)
-            if not movies:
-                continue
-            if self._enriched_count == 0:
-                self._enrich_start = time.perf_counter()
-            aggregated = await self._enrich_movies(movies)
-            self._enriched_count += len(aggregated)
-            self._log_progress(
-                "Enrichment",
-                self._enriched_count,
-                self._enrich_start,
-                self._points_queue.qsize(),
-            )
-            await self._emit_points(aggregated)
-
-    async def _process_episode_batch(self, batch: _EpisodeBatch) -> None:
-        for chunk in _chunk_sequence(batch.episodes, self._enrichment_batch_size):
-            episodes = list(chunk)
-            if not episodes:
-                continue
-            if self._enriched_count == 0:
-                self._enrich_start = time.perf_counter()
-            aggregated = await self._enrich_episodes(batch.show, episodes)
-            self._enriched_count += len(aggregated)
-            self._log_progress(
-                "Enrichment",
-                self._enriched_count,
-                self._enrich_start,
-                self._points_queue.qsize(),
-            )
-            await self._emit_points(aggregated)
-
-    async def _process_sample_batch(self, batch: _SampleBatch) -> None:
-        for chunk in _chunk_sequence(batch.items, self._enrichment_batch_size):
-            aggregated = list(chunk)
-            if not aggregated:
-                continue
-            if self._enriched_count == 0:
-                self._enrich_start = time.perf_counter()
-            self._enriched_count += len(aggregated)
-            self._log_progress(
-                "Enrichment",
-                self._enriched_count,
-                self._enrich_start,
-                self._points_queue.qsize(),
-            )
-            await self._emit_points(aggregated)
-
-    async def _enrich_movies(
-        self, movies: Sequence[PlexPartialObject]
-    ) -> list[AggregatedItem]:
-        movie_ids = [_extract_external_ids(movie) for movie in movies]
-        imdb_ids = [ids.imdb for ids in movie_ids if ids.imdb]
-        imdb_map = (
-            await _fetch_imdb_batch(self._http_client, imdb_ids)
-            if imdb_ids
-            else {}
-        )
-
-        api_key = self._tmdb_api_key
-        tmdb_results: list[TMDBMovie | None] = []
-        if api_key:
-            tmdb_tasks = [
-                _fetch_tmdb_movie(self._http_client, ids.tmdb, api_key)
-                for ids in movie_ids
-                if ids.tmdb
-            ]
-            if tmdb_tasks:
-                tmdb_results = await asyncio.gather(*tmdb_tasks)
-        tmdb_iter = iter(tmdb_results)
-
-        aggregated: list[AggregatedItem] = []
-        for movie, ids in zip(movies, movie_ids):
-            tmdb = next(tmdb_iter, None) if ids.tmdb else None
-            imdb = imdb_map.get(ids.imdb) if ids.imdb else None
-            aggregated.append(
-                AggregatedItem(
-                    plex=_build_plex_item(movie),
-                    imdb=imdb,
-                    tmdb=tmdb,
-                )
-            )
-        return aggregated
-
-    async def _enrich_episodes(
-        self, show: PlexPartialObject, episodes: Sequence[PlexPartialObject]
-    ) -> list[AggregatedItem]:
-        show_ids = _extract_external_ids(show)
-        show_tmdb: TMDBShow | None = None
-        if show_ids.tmdb:
-            show_tmdb = await self._get_tmdb_show(show_ids.tmdb)
-        episode_ids = [_extract_external_ids(ep) for ep in episodes]
-        imdb_ids = [ids.imdb for ids in episode_ids if ids.imdb]
-        imdb_map = (
-            await _fetch_imdb_batch(self._http_client, imdb_ids)
-            if imdb_ids
-            else {}
-        )
-
-        tmdb_results: list[TMDBEpisode | None] = []
-        if show_tmdb:
-            episode_tasks = [
-                self._lookup_tmdb_episode(show_tmdb, ep)
-                for ep in episodes
-            ]
-            if episode_tasks:
-                tmdb_results = await asyncio.gather(*episode_tasks)
-        tmdb_iter = iter(tmdb_results)
-        aggregated: list[AggregatedItem] = []
-        for ep, ids in zip(episodes, episode_ids):
-            tmdb_episode = next(tmdb_iter, None) if show_tmdb else None
-            imdb = imdb_map.get(ids.imdb) if ids.imdb else None
-            tmdb_item: TMDBItem | None = tmdb_episode or show_tmdb
-            aggregated.append(
-                AggregatedItem(
-                    plex=_build_plex_item(ep),
-                    imdb=imdb,
-                    tmdb=tmdb_item,
-                )
-            )
-        return aggregated
-
-    async def _get_tmdb_show(self, tmdb_id: str) -> TMDBShow | None:
-        if tmdb_id in self._show_tmdb_cache:
-            return self._show_tmdb_cache[tmdb_id]
-        show = await _fetch_tmdb_show(self._http_client, tmdb_id, self._tmdb_api_key or "")
-        self._show_tmdb_cache[tmdb_id] = show
-        return show
-
-    async def _lookup_tmdb_episode(
-        self, show_tmdb: TMDBShow | None, episode: PlexPartialObject
-    ) -> TMDBEpisode | None:
-        if not show_tmdb or not self._tmdb_api_key:
-            return None
-        season = resolve_tmdb_season_number(show_tmdb, episode)
-        ep_num = getattr(episode, "index", None)
-        if isinstance(ep_num, str) and ep_num.isdigit():
-            ep_num = int(ep_num)
-        if season is None or ep_num is None:
-            return None
-        return await _fetch_tmdb_episode(
-            self._http_client,
-            show_tmdb.id,
-            season,
-            ep_num,
-            self._tmdb_api_key,
-        )
-
-    async def _emit_points(self, aggregated: Sequence[AggregatedItem]) -> None:
-        if not aggregated:
-            return
-        self._items.extend(aggregated)
-        points = [
-            build_point(item, self._dense_model_name, self._sparse_model_name)
-            for item in aggregated
-        ]
-        await self._persistence_stage.enqueue_points(points)
-
-    async def _perform_upsert(
-        self, batch: Sequence[models.PointStruct]
-    ) -> None:
-        await _upsert_in_batches(
-            self._client,
-            self._collection_name,
-            list(batch),
-            retry_queue=self._qdrant_retry_queue,
-        )
-
-    def _handle_upsert_batch(
-        self, worker_id: int, batch_size: int, queue_size: int
-    ) -> None:
-        if self._upserted_points == 0:
-            self._upsert_start = time.perf_counter()
-        self._upserted_points += batch_size
-        self._log_progress(
-            f"Upsert worker {worker_id}",
-            self._upserted_points,
-            self._upsert_start,
-            queue_size,
-        )
 
 
 def _build_loader_orchestrator(
